inferences/requirements.txt
# Core dependencies for crack detection inference
mmcv>=2.0.0
mmsegmentation>=1.0.0
mmengine>=0.8.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-image>=0.19.0
slidingwindow>=0.0.13
Pillow>=8.0.0
matplotlib>=3.5.0

